Commit Subject: Rename files and fix errors in multiple modules
Commit Description:
