+++ .gitignore
@@ -7,12 +7,12 @@ __pycache__/
 .env
 /.env
 /tests/test_load/bloom_filter_state.bin
-src/monkey_patch/__pycache__/*
-src/monkey_patch/trackers/__pycache__/*
-src/monkey_patch/models/__pycache__/*
-src/monkey_patch/language_models/__pycache__/*
+src/tanuki/__pycache__/*
+src/tanuki/trackers/__pycache__/*
+src/tanuki/models/__pycache__/*
+src/tanuki/language_models/__pycache__/*
 src/__pycache__/*
 **/__pycache__/
 src/trackers/venv/
-/src/monkey_patch.py.egg-info/
+/src/tanuki.py.egg-info/
 /dist/
