+++ src/monkey_patch.py.egg-info/requires.txt
@@ -1,0 +1,1 @@
+tanuki.py
