+++ src/monkey_patch.py.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+tanuki
