+++ src/tanuki/aligns/align_classify_sentiment.py
@@ -3,7 +3,7 @@ import os
 import openai
 from dotenv import load_dotenv
 
-from monkey_patch.monkey import Monkey as monkey
+from tanuki.monkey import Monkey as tanuki
 
 import unittest
 from typing import Literal, Optional
@@ -14,19 +14,19 @@ openai.api_key = os.getenv("OPENAI_API_KEY")
 
 class TestClassifySentiment(unittest.TestCase):
 
-    @monkey.patch
+    @tanuki.patch
     def classify_sentiment_2(self, input: str, input_2: str) -> Optional[Literal['Good', 'Bad']]:
         """
         Determine if the inputs are positive or negative sentiment, or None
         """
 
-    @monkey.patch
+    @tanuki.patch
     def classify_sentiment(self, input: str) -> Optional[Literal['Good', 'Bad']]:
         """
         Determine if the input is positive or negative sentiment
         """
 
-    @monkey.align
+    @tanuki.align
     def test_align_classify_sentiment(self):
         """We can test the function as normal using Pytest or Unittest"""
