+++ tests/test_patch/test_finance.py
@@ -4,7 +4,7 @@ from typing import List
 import openai
 from dotenv import load_dotenv
 
-from monkey_patch.monkey import Monkey
+from tanuki.monkey import Monkey
 
 load_dotenv()
 openai.api_key = os.getenv("OPENAI_API_KEY")
